Added bootstrap heap of 2MB

+++ source/Debug.hpp
@@ -10,7 +10,7 @@ inline void PANIC(const char* str)
   while (true);
 }
 
-inline void debug(const char* str, int val)
+inline void debug(const char* str, long val)
 {
   Screen::putString(str);
   Screen::putString(" ");
